Conductor: fixes after merging updated staging
