+++ code/go/0chain.net/conductor/conductor/executor.go
@@ -772,7 +772,9 @@ func (r *Runner) ConfigureTestCase(configurator cases.TestCaseConfigurator) erro
 			state.SendInsufficientProposals = cfg
 
 		case *cases.VerifyingNonExistentBlock:
-			state.VerifyingNonExistentBlock = cfg
+			if state.IsMonitor {
+				state.VerifyingNonExistentBlock = cfg
+			}
 
 		case *cases.NotarisingNonExistentBlock:
 			state.NotarisingNonExistentBlock = cfg
