+++ docker.local/config/conductor.no-view-change.byzantine.yaml
@@ -85,7 +85,7 @@ tests:
 
   - name: "verifying non existent block"
     flow:
-      - set_monitor: "sharder-1"
+      - set_monitor: "miner-1"
       - cleanup_bc: {}
       - start_lock: [ "miner-1", "miner-2", "miner-3", "miner-4", "sharder-1", "sharder-2"]
       - configure_verifying_non_existent_block_test_case:
